F6.2 improve error messages
The warning "Footprint centers of pads and F.Fab do not match" was not
usefull. It could be understood as 'this needs to be fixed'. Also there is
no real use-case where this warning actually helps the user.

Instead now one line for each check is output which shows the calculated
centers. This way easier to read than the previous solution which had all
the coordinates in the same line.